move geolocation filter tests to filter module
