+++ tests/src/mod.rs
@@ -17,6 +17,5 @@ mod kv;
 mod query;
 mod scan;
 mod udf;
-mod geo;
 mod cdt_list;
 mod cdt_map;
